reorganize watcher modules for more consistency
