+++ crates/gitbutler-watcher/src/handler.rs
