+++ crates/gitbutler-watcher/src/lib.rs
@@ -19,8 +19,6 @@ use tokio::{
 use tokio_util::sync::CancellationToken;
 
 mod debouncer;
-mod debouncer_cache;
-mod debouncer_event;
 mod file_monitor;
 mod handler;
